+++ app/utils.js
@@ -411,7 +411,7 @@ function getTxTotalInputOutputValues(tx, txInputs, blockHeight) {
 
 				if (txInput) {
 					try {
-						var vout = txInput.vout[tx.vin[i].vout];
+						var vout = txInput;
 						if (vout.value) {
 							totalInputValue = totalInputValue.plus(new Decimal(vout.value));
 						}